refactor(product-pages): extract shared enroll primitives (types, EnrollButton, BoxGrid)
