+++ frontends/main/src/app-pages/ProductPages/useCourseEnrollment.ts
@@ -1,4 +1,3 @@
-import React from "react"
 import { useQuery } from "@tanstack/react-query"
 import type {
   CourseRunV2,
@@ -17,23 +16,17 @@ import { env } from "@/env"
 import { DASHBOARD_HOME } from "@/common/urls"
 import { getCourseScenario, type CourseScenario } from "./courseRun"
 import { useCourseEnrolledRunIds } from "./useCourseEnrolledRunIds"
+import type {
+  EnrollAction,
+  EnrollActionKind,
+  EnrollAreaState,
+} from "./enrollTypes"
 
-// "free" covers every non-paid enrollment — active audit ("Start Learning") and
-// the degraded archived/deadline-passed audit ("Access Course Materials"). They
-// hit the same free-enrollment path; only the button label differs.
-export type EnrollActionKind = "paid" | "free"
-
-export type EnrollAction = {
-  kind: EnrollActionKind
-  label: string
-  onClick?: (e: React.MouseEvent<HTMLButtonElement>) => void
-}
-
-// Discriminated — "enrolled" is the collapse state, not an "option"; "none" = no button.
-export type EnrollAreaState =
-  | { status: "enrolled"; href: string }
-  | { status: "options"; options: EnrollAction[] }
-  | { status: "none" }
+export type {
+  EnrollAction,
+  EnrollActionKind,
+  EnrollAreaState,
+} from "./enrollTypes"
 
 export type UseCourseEnrollment = {
   state: EnrollAreaState
